requirements.txt
openai
faiss-cpu
numpy
python-dotenv
tenacity
PyPDF2
pydantic